Refactor: Organize Apify Instagram scraping modules into comments and posts
