+++ lib/apify/comments/saveApifyInstagramComments.ts
@@ -1,8 +1,8 @@
 import { insertPostComments } from "@/lib/supabase/post_comments/insertPostComments";
-import getOrCreatePostsForComments from "@/lib/apify/getOrCreatePostsForComments";
-import getOrCreateSocialsForComments from "@/lib/apify/getOrCreateSocialsForComments";
+import getOrCreatePostsForComments from "@/lib/apify/comments/getOrCreatePostsForComments";
+import getOrCreateSocialsForComments from "@/lib/apify/comments/getOrCreateSocialsForComments";
 import { TablesInsert } from "@/types/database.types";
-import { InstagramComment } from "@/lib/apify/handleInstagramCommentsScraper";
+import { InstagramComment } from "@/lib/apify/comments/handleInstagramCommentsScraper";
 
 /**
  * Saves Instagram comments to the post_comments table in the database
