+++ lib/apify/posts/handleInstagramProfileFollowUpRuns.ts
@@ -1,5 +1,5 @@
-import runInstagramCommentsScraper from "@/lib/apify/runInstagramCommentsScraper";
-import getExistingPostComments from "@/lib/apify/getExistingPostComments";
+import runInstagramCommentsScraper from "@/lib/apify/comments/runInstagramCommentsScraper";
+import getExistingPostComments from "@/lib/apify/comments/getExistingPostComments";
 import { ApifyInstagramPost, ApifyInstagramProfileResult } from "@/types/Apify";
 
 /**
